Confirm-ul pentru NewEntry nu merge

+++ Model/Event.cs
@@ -3,7 +3,7 @@
 namespace mpp_proiect_csharp_DianaGliga11.Model
 {
 
-    public class Event : Entity<int>
+    public class Event : Entity<long>
     {
         public int Distance { get; set; }
